+++ test/formatters/samplecode/treefmt/out.rs
@@ -1,0 +1,9 @@
+fn foo() {
+    println!("a");
+}
+
+fn bar() {
+    println!("b");
+
+    println!("c");
+}